+++ test/client/errors/RequestError.js
@@ -1,11 +1,17 @@
 const {RequestError, HaystackError, GraphQLError} = require("../../../src/errors");
 const { expect } = require("chai");
+const {StubLogger} = require("../../stubs");
 
 describe("RequestError", () => {
+    let logger = new StubLogger();
+    beforeEach(() => {
+        logger.resetHistory();
+    });
+
     describe("make", () => {
         it("should not create RequestError if no response received", () => {
             const error = new Error("Test error");
-            const reqError = RequestError.make(error);
+            const reqError = RequestError.make(error, logger);
             expect(reqError).to.be.instanceof(Error);
         });
 
@@ -13,7 +19,7 @@ describe("RequestError", () => {
             // e.g. HTTP 403 errors don't typically have a message
             const error = new Error("Test error");
             error.response = {};
-            const reqError = RequestError.make(error);
+            const reqError = RequestError.make(error, logger);
             expect(reqError).to.be.instanceof(Error);
         });
 
@@ -24,7 +30,7 @@ describe("RequestError", () => {
                     nothingToSee: "here"
                 }
             };
-            const reqError = RequestError.make(error);
+            const reqError = RequestError.make(error, logger);
             expect(reqError).to.be.instanceof(Error);
         });
 
@@ -38,7 +44,7 @@ describe("RequestError", () => {
                         }
                     }
                 };
-                const reqError = RequestError.make(error);
+                const reqError = RequestError.make(error, logger);
                 expect(reqError).to.be.instanceof(HaystackError);
                 expect(reqError.message).to.equal("HBadRequestError: It broken");
             });
@@ -63,12 +69,14 @@ describe("RequestError", () => {
                         ]
                     }
                 };
-                const reqError = RequestError.make(error);
+                const reqError = RequestError.make(error, logger);
                 expect(reqError).to.be.instanceof(GraphQLError);
                 expect(reqError.message).to.equal(
                     "Field \"search\" argument \"whereTag\" of type \"String!\" is required but not provided."
                 );
-                expect(reqError.errors).to.eql(error.response.data.errors);
+                expect(reqError.errors).to.eql([
+                    "Field \"search\" argument \"whereTag\" of type \"String!\" is required but not provided. @ location/s line 5:5"
+                ]);
             });
         });
 
@@ -99,10 +107,13 @@ describe("RequestError", () => {
                     ]
                 }
             };
-            const reqError = RequestError.make(error);
+            const reqError = RequestError.make(error, logger);
             expect(reqError).to.be.instanceof(GraphQLError);
-            expect(reqError.message).to.equal("More than 1 error encountered");
-            expect(reqError.errors).to.eql(error.response.data.errors);
+            expect(reqError.message).to.equal("More than 1 GraphQLError encountered");
+            expect(reqError.errors).to.eql([
+                "Field \"search\" argument \"whereTag\" of type \"String!\" is required but not provided. @ location/s line 5:5",
+                "Field blah not blah blah @ location/s line 5:10"
+            ]);
         });
     });
 });